+++ R/create.R
@@ -1,102 +1,142 @@
-#' Two-mode lattice
+#' Two-mode chain
 #'
-#' Creates a two-mode lattice
-#' @param node1 Number of nodes in the first node set
-#' @param node2 Number of nodes in the second node set
+#' Creates a two-mode chain
+#' @param n1 Number of nodes in the first node set
+#' @param n2 Number of nodes in the second node set
+#' @param as What type of object to return.
+#' One of "matrix", "tbl_graph", "igraph".
+#' By default, creates a "tbl_graph" object.
 #' @details Will construct a bilateral lattice,
 #' with two ties for every second-mode node.
 #' @export
 #' @examples
 #' \dontrun{
-#' create_lattice(10, 12)
+#' create_chain(5,10) %>% ggraph() +
+#' geom_edge_fan(aes(alpha = stat(index)), show.legend = FALSE) +
+#' geom_node_point(aes(size = 5))
 #' }
-create_lattice <- function(node1, node2) {
-  mat <- matrix(0, node1, node2)
+create_chain <- function(n1, n2, 
+                         as = c("tbl_graph", "igraph", "matrix")) {
+  
+  mat <- matrix(0, n1, n2)
   out <- suppressWarnings(((row(mat) - col(mat)) == 0 |
-    (row(mat) - col(mat)) == (-seq.int(0, node2 - 1, node1)[-1]) |
+    (row(mat) - col(mat)) == (-seq.int(0, n2 - 1, n1)[-1]) |
     (row(mat) - col(mat)) == -1 |
-    (row(mat) - col(mat)) == -seq.int(1 + node1, node2 - 1, node1) |
+    (row(mat) - col(mat)) == -seq.int(1 + n1, n2 - 1, n1) |
     (row(mat) - col(mat)) == nrow(mat) - 1) * 1)
+  
+  as <- match.arg(as)
+  if(as == "tbl_graph") out <- tidygraph::as_tbl_graph(out)
+  if(as == "igraph") out <- igraph::graph_from_incidence_matrix(out)
   out
-}
-
-#' Two-mode random graph
-#'
-#' Creates a random two-mode network
-#' @param node1 Number of nodes in the first node set
-#' @param node2 Number of nodes in the second node set
-#' @param density Number of edges in the network over the number of edges possible
-#' @details Will construct an affiliation matrix,
-#' with a random probability of a tie.
-#' @export
-#' @examples
-#' \dontrun{
-#' create_random(10, 12, 0.25)
-#' }
-create_random <- function(node1, node2, density) {
-  mat <- matrix(rbinom(node1 * node2, 1, density), node1, node2)
-  mat
-}
-
-#' Two-component two-mode graph
-#'
-#' Creates a two-component two-mode network
-#' @param node1 Number of nodes in the first node set
-#' @param node2 Number of nodes in the second node set
-#' @details Will construct an affiliation matrix,
-#' with full component diagonal.
-#' @export
-#' @examples
-#' \dontrun{
-#' create_poles(10, 12)
-#' }
-create_poles <- function(node1, node2) {
-  mat <- matrix(0, node1, node2)
-  mat[1:round(node1 / 2), 1:round(node2 / 2)] <- 1
-  mat[rowSums(mat) < 1, colSums(mat) < 1] <- 1
-  mat
 }
 
 #' Matched two-mode graph
 #'
 #' Creates a matched two-mode network
-#' @param node1 Number of nodes in the first node set
-#' @param node2 Number of nodes in the second node set
+#' @param n1 Number of nodes in the first node set
+#' @param n2 Number of nodes in the second node set
 #' @details Will construct an affiliation matrix,
-#' with by default both node1 and node2 matched.
-#' TODO: Make node set matching optional.
+#' with by default both n1 and n2 matched.
+#' TODO: Incorporate into create_chain (chordal_ring of certain breadth w).
 #' @export
 #' @examples
 #' \dontrun{
 #' create_match(10, 12)
 #' }
-create_match <- function(node1, node2) {
-  mat <- matrix(0, node1, node2)
+create_match <- function(n1, n2,
+                         as = c("tbl_graph", "igraph", "matrix")) {
+  mat <- matrix(0, n1, n2)
   mat <- matrix(
-    ((row(mat) - col(mat)) %in% unique(-seq(0, node2 - 1, node1), 
-                                       seq(0, node1 - 1, node2))) * 1,
-    node1, node2
+    ((row(mat) - col(mat)) %in% unique(-seq(0, n2 - 1, n1), 
+                                       seq(0, n1 - 1, n2))) * 1,
+    n1, n2
   )
+
+  as <- match.arg(as)
+  if(as == "tbl_graph") mat <- tidygraph::as_tbl_graph(mat)
+  if(as == "igraph") mat <- igraph::graph_from_incidence_matrix(mat)
+  mat
+}
+
+#' Two-mode random graph
+#'
+#' Creates a random two-mode network
+#' @param n1 Number of nodes in the first node set
+#' @param n2 Number of nodes in the second node set
+#' @param p Number of edges in the network over the number of edges possible
+#' @param m Number of edges in the network
+#' @param as What type of object to return.
+#' One of "matrix", "tbl_graph", "igraph".
+#' By default, creates a "tbl_graph" object.
+#' @details Will construct an affiliation matrix,
+#' with a random probability of a tie.
+#' @export
+#' @examples
+#' \dontrun{
+#' play_twomode(10, 12, 0.25) %>% ggraph() +
+#' geom_edge_fan(aes(alpha = stat(index)), show.legend = FALSE) +
+#' geom_node_point(aes(size = 5))
+#' }
+play_twomode <- function(n1, n2, p, m, directed = TRUE, mode = "out",
+                           as = c("tbl_graph", "igraph", "matrix")) {
+  
+  g <- tidygraph::play_bipartite(n1, n2, p, m, directed, mode)
+
+  as <- match.arg(as)
+  if(as == "igraph" | as == "matrix") g <- as.igraph(g)
+  if(as == "matrix") g <- as_adjacency_matrix(g)
+  g
+}
+
+#' Two-component two-mode graph
+#'
+#' Creates a two-component two-mode network
+#' @param n1 Number of nodes in the first node set
+#' @param n2 Number of nodes in the second node set
+#' @details Will construct an affiliation matrix,
+#' with full component diagonal.
+#' TODO: Allow specfication of how many silos/components to create
+#' @export
+#' @examples
+#' \dontrun{
+#' create_silos(10, 12)
+#' }
+create_silos <- function(n1, n2,
+                         as = c("tbl_graph", "igraph", "matrix")) {
+  mat <- matrix(0, n1, n2)
+  mat[1:round(n1 / 2), 1:round(n2 / 2)] <- 1
+  mat[rowSums(mat) < 1, colSums(mat) < 1] <- 1
+
+  as <- match.arg(as)
+  if(as == "tbl_graph") mat <- tidygraph::as_tbl_graph(mat)
+  if(as == "igraph") mat <- igraph::graph_from_incidence_matrix(mat)
   mat
 }
 
 #' Nested two-mode graph
 #'
 #' Creates a nested two-mode network
-#' @param node1 Number of nodes in the first node set
-#' @param node2 Number of nodes in the second node set
+#' @param n1 Number of nodes in the first node set
+#' @param n2 Number of nodes in the second node set
 #' @details Will construct an affiliation matrix,
-#' with decreasing fill across node2.
-#' TODO: Make node set matching optional.
+#' with decreasing fill across n2.
 #' @export
 #' @examples
 #' \dontrun{
 #' create_nest(10, 12)
 #' }
-create_nest <- function(node1, node2) {
-  mat <- matrix(0, node1, node2)
-  mat[(row(mat) - col(mat)) >= 0] <- 1
-  mat
+create_nest <- function(n1, n2, 
+                        as = c("tbl_graph", "igraph", "matrix")) {
+  
+  as <- match.arg(as)
+  
+  out <- matrix(0, n1, n2)
+  out[(row(out) - col(out)) >= 0] <- 1
+
+  if(as == "tbl_graph") out <- tidygraph::as_tbl_graph(out)
+  if(as == "igraph") out <- igraph::graph_from_incidence_matrix(out)
+  out
 }
 
 # mat.dist <- matrix(0,5,3)